% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alarm.R
\name{get_alarm}
\alias{get_alarm}
\title{Get inverter device alarms}
\usage{
get_alarm(device_ip, model = "APSystems")
}
\arguments{
\item{device_ip}{list or vector of devices IP address}

\item{model}{the inverter device model. Currently only "APSystems" is supported.}
}
\value{
a dataframe with one row of device information per `device_id` answering the query.
}
\description{
Get inverter device alarms
}
\examples{
\dontrun{
get_alarm(c("192.168.0.12", "192.168.0.230"))
}
}
