% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_device_info}
\alias{get_device_info}
\title{Get inverter device information}
\usage{
get_device_info(device_ip, model = "APSystems")
}
\arguments{
\item{device_ip}{list or vector of APSystem devices IP addresses}

\item{model}{the inverter device model. Currrently only "APSystems" is supported.}
}
\value{
a dataframe with one row of device information per `device_id` answering the query.
}
\description{
Get inverter device information
}
\examples{
get_device_info(c("192.168.0.12", "192.168.0.230"))
}
