% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_device.R
\name{query_ap_device}
\alias{query_ap_device}
\title{AP System single device query}
\usage{
query_ap_device(device_ip, query)
}
\arguments{
\item{device_ip}{IP address of the APSystem device}

\item{query}{the API query string}
}
\value{
a data-frame with a `device_id` column and the `$data` turned into
   as many columns as expected
}
\description{
AP System single device query
}
\examples{
\dontrun{
query_ap_device(device_ip = "192.168.0.234", query = "getDeviceInfo")
}
}
