% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_device.R
\name{query_ap_devices}
\alias{query_ap_devices}
\title{AP System multi-device query}
\usage{
query_ap_devices(device_ip, query)
}
\arguments{
\item{device_ip}{list or vector of each APSystem device IP address}

\item{query}{the API query string}
}
\value{
a data-frame with a row for each `device_id`, and the `$data` turned into
   as many columns as expected
}
\description{
AP System multi-device query
}
\examples{
\dontrun{
query_ap_devices(device_ip = c("192.168.0.234", "192.168.0.235"),
                 query = "getDeviceInfo"
                 )
}
}
